refactor: took out clients into seperate module
fixed a lil documentation too
